+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Check userAgent for SSR
 
 ## [1.0.5] - 2020-02-03
 ### Changed